+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.4.0] - Unrelease
 
 - [DemoApp/SDK] #159 Add ODSSpacer component and guidelines spacing
+- [DemoApp/SDK] #36 Add List component
 - [Component] #166 Add Text Selection component and remove Text Edit Menu
 - [Component] #35 Add tab bar page
 - [DemoApp] Bug #155 Typography style addition, and spacing fix